// asb-good-foods-user/src/app/common/models/appointment.model.ts
export class Appointment {
    eventType:string;
    eventName: string;
    eventDate: any;
    eventTime: any;
    noOfSlots: number;
    items : any;
    location : string;
    fullName: string;
    mobile: number;
    email: string;
    addressLine1: string;
    addressLine2: string;
    city: string;
    state: string;
    pincode: number;
    eventId:string;
    appointmentDocId:string;
    userUid: string;
}